Import of EXIFTOOL/Image-ExifTool-8.10 from CPAN.
gitpan-cpan-distribution: Image-ExifTool
gitpan-cpan-version:      8.10
gitpan-cpan-path:         EXIFTOOL/Image-ExifTool-8.10.tar.gz
gitpan-cpan-author:       EXIFTOOL
gitpan-cpan-maturity:     released

+++ t/MWG.t
@@ -1,11 +1,7 @@
-# Before `make install' is performed this script should be runnable with
-# `make test'. After `make install' it should work as `perl t/MWG.t'
+# Before "make install", this script should be runnable with "make test".
+# After "make install" it should work as "perl t/MWG.t".
 
-######################### We start with some black magic to print on failure.
-
-# Change "1..N" below to so that N matches last test number
-
-BEGIN { $| = 1; print "1..4\n"; $Image::ExifTool::noConfig = 1; }
+BEGIN { $| = 1; print "1..5\n"; $Image::ExifTool::noConfig = 1; }
 END {print "not ok 1\n" unless $loaded;}
 
 # test 1: Load the module(s)
@@ -13,8 +9,6 @@ use Image::ExifTool 'ImageInfo';
 use Image::ExifTool::MWG;
 $loaded = 1;
 print "ok 1\n";
-
-######################### End of black magic.
 
 use t::TestLib;
 
@@ -48,14 +42,29 @@ my $testnum = 1;
         $testfile = "t/${testname}_${testnum}_failed.jpg";
         unlink $testfile;
         $exifTool->WriteInfo("t/images/$src", $testfile);
-        my $info = $exifTool->ImageInfo($testfile, @tags);
-        if (check($exifTool, $info, $testname, $testnum)) {
-            unlink $testfile;
-        } else {
+        my $info = $exifTool->GetInfo('Warning');
+        if ($$info{Warning}) {
+            warn "\n    Warning: $$info{Warning}\n";
             print 'not ';
+        } else {
+            $info = $exifTool->ImageInfo($testfile, @tags);
+            if (check($exifTool, $info, $testname, $testnum)) {
+                unlink $testfile;
+            } else {
+                print 'not ';
+            }
         }
         print "ok $testnum\n";
     }
+}
+
+# test 5: Extract IPTC information from non-standard image while in strict MWG mode
+{
+    ++$testnum;
+    my $exifTool = new Image::ExifTool;
+    my $info = $exifTool->ImageInfo('t/images/ExifTool.jpg', 'IPTC:*', 'Warning');
+    print 'not ' unless check($exifTool, $info, $testname, $testnum);
+    print "ok $testnum\n";
 }
